requirements.txt
asgiref==3.3.4
click==8.0.1
fastapi==0.65.2
h11==0.12.0
pydantic==1.8.2
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
